Issue #29: Fixed camera model for left and right, projection utils, interlacing and save to file.
It was discovered that by calling set_projection_matrix which sets UseExplicitProjectionTransformMatrixOn
on the vtkCamera, this broke the ability to use vtkWindowToImageFilter.

So, this large change required me to setup the vtkCamera directly using benoitrosa's method.
Then projection, overlay and interlacing can all work, even when the window is resized to arbitrary sizes.